requirements.txt
Bootstrap-Flask==1.5.1
click==7.1.2
dominate==2.6.0
Flask==1.1.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.15.0
tencentcloud-sdk-python==3.0.351
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
